Cdp async test

+++ includes/class-wp-es-feeder.php
@@ -24,6 +24,7 @@ if ( !class_exists( 'wp_es_feeder' ) ) {
     }
 
     private function load_dependencies() {
+      require_once plugin_dir_path( dirname( __FILE__ ) ) . 'vendor/autoload.php';
       require_once plugin_dir_path( dirname( __FILE__ ) ) . 'includes/class-wp-es-feeder-loader.php';
       require_once plugin_dir_path( dirname( __FILE__ ) ) . 'admin/class-wp-es-feeder-admin.php';
       $this->loader = new wp_es_feeder_Loader();
@@ -52,6 +53,9 @@ if ( !class_exists( 'wp_es_feeder' ) ) {
       add_action( 'delete_post', array( &$this, 'delete_post' ), 10, 1 );
       add_action( 'trash_post', array( &$this, 'delete_post' ) );
       add_action( 'wp_ajax_es_request', array( $this, 'es_request') );
+      add_action( 'wp_ajax_es_sync_status', array($this, 'get_sync_status') );
+      add_action( 'wp_ajax_es_initiate_sync', array($this, 'es_initiate_sync') );
+      add_action( 'wp_ajax_es_process_next', array($this, 'es_process_next') );
     }
 
     public function run() {
@@ -74,15 +78,77 @@ if ( !class_exists( 'wp_es_feeder' ) ) {
       return $this->proxy;
     }
 
+    /**
+     * Triggered via AJAX, retreives a post's sync status for display.
+     */
+    public function get_sync_status() {
+      $post_id = $_POST['post_id'];
+      $status = get_post_meta($post_id, '_cdp_sync_status', true);
+      echo ES_FEEDER_SYNC::display($status);
+      exit;
+    }
+
+    /**
+     * Triggered via AJAX, clears out old sync data and initiates a new sync process.
+     */
+    public function es_initiate_sync() {
+      check_admin_referer();
+      global $wpdb;
+      $wpdb->delete($wpdb->postmeta, array('meta_value' => '_cdp_sync_queue'));
+      $opts = get_option( $this->plugin_name );
+      $post_types = $opts[ 'es_post_types' ];
+      $formats = implode(',', array_fill(0, count($post_types), '%s'));
+      $query = "SELECT p.ID FROM $wpdb->posts p 
+                  LEFT JOIN (SELECT post_id, meta_value FROM $wpdb->postmeta WHERE meta_key = '_iip_index_post_to_cdp_option') m
+                    ON p.ID = m.post_id 
+                  WHERE p.post_type IN ($formats) AND p.post_status = 'publish' AND (m.meta_value IS NULL OR m.meta_value != 'no')";
+      $query = $wpdb->prepare($query, array_keys($post_types));
+      $post_ids = $wpdb->get_col($query);
+      if (!count($post_ids)) {
+        echo json_encode(array('error' => true, 'message' => 'No posts found.', 'query' => $query));
+        exit;
+      }
+      foreach ($post_ids as $post_id)
+        update_post_meta($post_id, '_cdp_sync_queue', 1);
+      $this->es_process_next();
+    }
+
+    /**
+     * Grabs the next post in the queue and sends it to the API.
+     * Updates the postmeta indicating that this post has been synced.
+     * Returns a JSON object containing the API response for the current post
+     * as well as stats on the sync queue.
+     */
+    public function es_process_next() {
+      check_admin_referer();
+      global $wpdb;
+      $query = "SELECT post_id FROM $wpdb->postmeta WHERE meta_key = '_cdp_sync_queue' AND meta_value = 1";
+      $post_id = $wpdb->get_var($query);
+      if (!$post_id) {
+        $query = "SELECT COUNT(*) as total, SUM(meta_value) as incomplete FROM $wpdb->postmeta WHERE meta_key = '_cdp_sync_queue'";
+        $row = $wpdb->get_row($query);
+        $wpdb->delete($wpdb->postmeta, array('meta_key' => '_cdp_sync_queue'));
+        echo json_encode(array('done' => 1, 'total' => $row->total, 'complete' => ($row->total - $row->incomplete)));
+        exit;
+      }
+      update_post_meta($post_id, '_cdp_sync_queue', "0");
+      $post = get_post($post_id);
+      $resp = $this->addOrUpdate($post, false);
+      $query = "SELECT COUNT(*) as total, SUM(meta_value) as incomplete FROM $wpdb->postmeta WHERE meta_key = '_cdp_sync_queue'";
+      $row = $wpdb->get_row($query);
+      echo json_encode(array('done' => 0, 'response' => $resp, 'total' => $row->total, 'complete' => $row->total - $row->incomplete));
+      exit;
+    }
+
     public function save_post( $id, $post ) {
       $settings  = get_option( $this->plugin_name );
       $post_type = $post->post_type;
 
       if (array_key_exists('index_post_to_cdp_option', $_POST)) {
         update_post_meta(
-            $id,
-            '_iip_index_post_to_cdp_option',
-            $_POST['index_post_to_cdp_option']
+          $id,
+          '_iip_index_post_to_cdp_option',
+          $_POST[ 'index_post_to_cdp_option' ]
         );
       }
 
@@ -91,6 +157,7 @@ if ( !class_exists( 'wp_es_feeder' ) ) {
         return;
       }
 
+
       // switch operation based on post status
       if ( $post->post_status === 'publish' ) {
       
@@ -98,7 +165,7 @@ if ( !class_exists( 'wp_es_feeder' ) ) {
         $shouldIndex = $_POST['index_post_to_cdp_option'];
         
         // default to indexing - post has to be specifically set to 'no'
-        if( $shouldIndex === 'no' ) { 
+        if( $shouldIndex === 'no' ) {
           $this->delete( $post );
         } else {
           $this->addOrUpdate( $post );
@@ -125,50 +192,104 @@ if ( !class_exists( 'wp_es_feeder' ) ) {
       $this->delete( $post );
     }
 
-    public function addOrUpdate( $post ) {
+    /**
+     * Determines if a post can be synced or not. Syncable means that it is not in the process
+     * of being synced. If it is not syncable, update the sync status to inform the user that
+     * they needs to wait until the sync is complete and then resync.
+     *
+     * @param $post
+     * @return bool
+     */
+    public function is_syncable( $post ) {
+      // check sync status
+      $sync_status = get_post_meta($post->ID, '_cdp_sync_status', true);
+      if (!ES_FEEDER_SYNC::sync_allowed($sync_status)) {
+        update_post_meta($post->ID, '_cdp_sync_status', ES_FEEDER_SYNC::SYNC_WHILE_SYNCING);
+        return false;
+      }
+      return true;
+    }
+
+    public function addOrUpdate( $post, $print = true ) {
+      if ( !$this->is_syncable( $post ) ) {
+        $response = ['error' => 1, 'message' => 'Could not sync while sync in progress.'];
+        if (!$print)
+          wp_send_json($response);
+        return $response;
+      }
+
       // plural form of post type
       $post_type_name = ES_API_HELPER::get_post_type_label( $post->post_type, 'name' );
       // settings and configuration for plugin
       $config = get_option( $this->plugin_name );
 
       // api endpoint for wp-json
-      $wp_api_url = '/elasticsearch/v1/'.rawurlencode($post_type_name).'/'.$post->ID;
+      $wp_api_url = '/'.ES_API_HELPER::NAME_SPACE.'/'.rawurlencode($post_type_name).'/'.$post->ID;
       $request = new WP_REST_Request('GET', $wp_api_url);
       $api_response = rest_do_request( $request );
       $api_response = $api_response->data;
 
-      if (!$api_response) {
+      if ( !$api_response || isset( $api_response[ 'code' ] ) ) {
         error_log( print_r( $this->error . 'addOrUpdate() calling wp rest failed', true ) );
-        return;
+        $api_response['error'] = true;
+        $api_response['url'] = $wp_api_url;
+        if ( $print ) {
+          wp_send_json( $api_response );
+        }
+        return $api_response;
       }
+
+      // create callback for this post
+      global $wpdb;
+      do {
+        $uid = uniqid();
+        $query = "SELECT post_id FROM $wpdb->postmeta WHERE meta_key = '_cdp_sync_uid' AND meta_value = '$uid'";
+      } while ($wpdb->get_var($query));
+      $callback = get_rest_url(null, ES_API_HELPER::NAME_SPACE . '/callback/' . $uid);
+      update_post_meta($post->ID, '_cdp_sync_uid', $uid);
+      update_post_meta($post->ID, '_cdp_sync_status', ES_FEEDER_SYNC::SYNCING);
 
       $options = array(
         'url' => $config[ 'es_url' ] . '/' . $post->post_type,
         'method' => 'POST',
-        'body' => $api_response
+        'body' => $api_response,
+        'print' => $print
       );
 
-      $response = $this->es_request( $options );
+      $response = $this->es_request( $options, $callback );
       if ( !$response ) {
         error_log( print_r( $this->error . 'addOrUpdate()[add] request failed', true ) );
       }
+
+      return $response;
     }
 
     public function delete( $post ) {
+      if ( !$this->is_syncable( $post ) ) return;
       $opt = get_option( $this->plugin_name );
 
       $uuid = $this->get_uuid($post);
-      $delete_url = $opt[ 'es_url' ] . '/' . $post->post_type . '/' . $uuid;
+      $delete_url = $opt[ 'es_url' ] . '/' . $post->post_type;
 
       $options = array(
          'url' => $delete_url,
-         'method' => 'DELETE'
+         'method' => 'DELETE',
+         'body' => array(
+           'post_id' => $post->ID,
+           'type' => $post->post_type,
+           'site' => $this->get_site()
+         ),
+         'print' => false
       );
 
-      $this->es_request( $options );
+      $response = $this->es_request( $options );
+      if (!isset($response['error']) || !$response['error']) {
+        update_post_meta( $post->ID, '_cdp_sync_status', ES_FEEDER_SYNC::NOT_SYNCED );
+        delete_post_meta( $post->ID, '_cdp_sync_uid' );
+      }
     }
 
-    public function es_request($request) {
+    public function es_request($request, $callback = null) {
       $is_internal = false;
       if (!$request) {
         $request = $_POST['data'];
@@ -176,62 +297,72 @@ if ( !class_exists( 'wp_es_feeder' ) ) {
         $is_internal = true;
       }
 
-      $curl = curl_init();
-      curl_setopt($curl, CURLOPT_URL, $request['url']);
-      curl_setopt($curl, CURLOPT_SSL_VERIFYHOST, 0);
-      curl_setopt($curl, CURLOPT_SSL_VERIFYPEER, 0);
-      curl_setopt($curl, CURLOPT_RETURNTRANSFER, 1);
-      curl_setopt($curl, CURLOPT_TIMEOUT, 10);
-      curl_setopt($curl, CURLOPT_CONNECTTIMEOUT, 10);
+      $headers = [];
+      if ($callback) $headers['callback'] = $callback;
 
-      // if a body is provided
-      if ($request['body']) {
-        // unwrap the post data from ajax call
-        if (!$is_internal) {
-          $body = urldecode(base64_decode($request['body']));
-        } else {
-          $body = json_encode($request['body']);
-        }
+      $client = new GuzzleHttp\Client();
 
-        // check if domain is mapped
-        $opt = get_option( $this->plugin_name );
-        $protocol = is_ssl() ? 'https://' : 'http://';
-        $opt_url = $opt['es_wpdomain'];
-        $opt_url = str_replace($protocol, '', $opt_url);
-        $site_url = site_url();
-        $site_url = str_replace($protocol, '', $site_url);
-
-        if ($opt_url !== $site_url) {
-          $body = str_replace($site_url, $opt_url, $body);
-        }
-
-        // obtain string length
-        $length = strlen($body);
-
-        // curl options
-        curl_setopt($curl, CURLOPT_CUSTOMREQUEST, $request['method']);
-        curl_setopt($curl, CURLOPT_POSTFIELDS, $body);
-        curl_setopt($curl, CURLOPT_HTTPHEADER, array(
-          'Content-Type: application/json',
-          'Content-Length: ' . $length
-        ));
-      } else {
-        curl_setopt($curl, CURLOPT_CUSTOMREQUEST, $request['method']);
-        curl_setopt($curl, CURLOPT_HTTPHEADER, array(
-          'Content-Type: application/json'
-        ));
+      // TODO: Investigate alternative methods to accomplish this URL check. The API gets mad since we don't have a GET route for this yet.
+      try {
+        $client->get($request['url'], ['http_errors' => false, 'timeout' => '5']);
+      } catch (GuzzleHttp\Exception\ConnectException $e) {
+        $error = json_encode($e->getHandlerContext());
       }
 
-      $results = curl_exec($curl);
-      curl_close($curl);
+      if ( isset($error) ){
+        $results = $error;
+      } else {
+        // if a body is provided
+        if ( isset($request['body']) ) {
+          // unwrap the post data from ajax call
+          if (!$is_internal) {
+            $body = urldecode(base64_decode($request['body']));
+          } else {
+            $body = json_encode($request['body']);
+            $headers['Content-Type'] = 'application/json';
+          }
 
-      if ($is_internal) {
+          $body = $this->is_domain_mapped($body);
+
+          $response = $client->request($request['method'], $request['url'], ['body' => $body, 'http_errors' => false, 'headers' => $headers]);
+        } else {
+          $response = $client->request($request['method'], $request['url'], ['http_errors' => false, 'headers' => $headers]);
+        }
+
+        $body = $response->getBody();
+        $results = $body->getContents();
+      }
+
+      if ($is_internal || (isset($request['print']) && !$request['print'])) {
         return json_decode($results);
       } else {
-        return wp_send_json(json_decode($results));
+        wp_send_json(json_decode($results));
+        return null;
       }
     }
 
+    private function is_domain_mapped( $body ) {
+      // check if domain is mapped
+      $opt = get_option( $this->plugin_name );
+      $protocol = is_ssl() ? 'https://' : 'http://';
+      $opt_url = $opt['es_wpdomain'];
+      $opt_url = str_replace($protocol, '', $opt_url);
+      $site_url = site_url();
+      $site_url = str_replace($protocol, '', $site_url);
+
+      if ($opt_url !== $site_url) {
+        $body = str_replace($site_url, $opt_url, $body);
+      }
+
+      return $body;
+    }
+
+    /**
+     * Construct UUID which is site domain delimited by dashes and not periods, underscore, and post ID.
+     *
+     * @param $post
+     * @return string
+     */
     public function get_uuid($post) {
       $opt = get_option( $this->plugin_name );
       $url = $opt['es_wpdomain'];
@@ -244,6 +375,18 @@ if ( !class_exists( 'wp_es_feeder' ) ) {
 
       $host = str_replace('.', '-', $host);
       return "{$host}_{$post->ID}";
+    }
+
+    public function get_site() {
+      $opt = get_option( ES_API_HELPER::PLUGIN_NAME );
+      $url = $opt[ 'es_wpdomain' ];
+      $args = parse_url( $url );
+      $host = $url;
+      if ( array_key_exists( 'host', $args ) )
+        $host = $args[ 'host' ];
+      else
+        $host = str_ireplace( 'https://', '', str_ireplace( 'http://', '', $host ) );
+      return $host;
     }
   }
 }
